Split keys features macros into their own file
